[Generator] Add first implementation of code generator
This is a simple, yet complete, steps code generator. The generated code
exploits the variadic template step wrapper provided by the
ProgramControlBlock class.

An example XML test program is added as a reference. This file is a simple
schema which is quite similar to the Doxygen generated description.

The code generator exploits the RapidXML template library which is provided
within the "include" folder.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>