[SyncBarrier] Add first implementation
This provides a simple implementation for a generic "Single-Producer Multiple-Consumer"
synchronization barrier. Such a barrier exposes the required support to
properly handle double-buffering of a generic datatype T; where T could be, for example,
a pointer to an image buffer.

This kind of barrier allows to setup both SPSC pipelines as well as SPMC
fork patterns. The only constraints are:
1. each consumer does not modify the input buffer;
2. each consumer does _at least one PULL_ of the input buffer;
3. both the producer and all the consumers uses (respectively) as output
   and input buffer :
   a. the one returned by buffer_push/buffer_pull
   b. the one returned by producerBuffer/consumerBuffer, after a push/pull
      operation has been performed.

TODO: optimize the template for the SPSC case in order to optimize
      pipeline based (sub)networks.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>