[PCB_DEBUG] Add StepOut support
The StepOut function allows to define where to stop in DEBUG mode, thus
allowing to execute up to a specified step and than block waiting for a
continue/repeat debug action.

To this purpose, the Step wrapper template has been extended to get a constant
non-type parameter which is used to identify the step being called and thus
making such an ID available to the PCB internal methods (i.e. CheckDebug) to
verify if the current step has been marked with a breakpoint (i.e. stepout).

Signed-off-by: Patrick Bellasi <derkling@gmail.com>